capsense: Handle unexpected reset of IQS231A controller.
Noise on the I2C bus can lead to an unexpected reset of the IQS231A controller
in which case the chip will go through a power-on-reset sequence, and requires
around 500ms of silence on the I2C bus in order to recover. This has been seen
on badges when handling them roughly or when taking less-than-perfect ESD
precautions.