Add BLE sources for badge and beacon firmware.
The BLE firmware is built using the Zephyr RTOS, and targets either
the Taiyo Yuden EYSGCNZWY or the Nordic nrf51822 beacon. The firmware
is mostly just used for exchanging emotes, starting the 'awoo' event
and for passing rabies on to other badges.

There were multiple releases of this firmware made during DEFCON, this
represents the final release, which should have made it on to
approximately half of the badges.